updated and rerouted section two naming conventions
